pyproject.toml
[project]
name = "software-builder"
version = "0.1.0"
description = "Local LLM coding agent with built-in evaluation system"
readme = "readme.md"
authors = [
    { name = "Ivan Willig", email = "iwillig@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    # Core
    "click>=8.1.0",
    "sqlalchemy>=2.0.0",
    "llm>=0.15.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    # Web UI (Phase 1)
    "flask>=3.0.0",
    "datastar-py>=0.3.0",
    # Content rendering (shared by both UIs)
    "markdown>=3.5.0",
    "markdown-full-yaml-metadata>=2.2.0",
    "pygments>=2.17.0",
    "alembic>=1.18.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-flask>=1.3.0",
]

[project.scripts]
software-builder = "software_builder.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Ruff configuration - linting and formatting
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
# Enable pycodestyle (E), Pyflakes (F), isort (I), and more
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings  
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Markers for different test types
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "eval: LLM evaluation tests",
    "slow: Slow-running tests",
]

# Coverage settings
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",  # Show summary of all test outcomes
]

# Logging
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# Coverage
[tool.coverage.run]
source = ["src/software_builder"]
omit = [
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
